tests/wpt/tests/close-watcher/user-activation/yy.html

<!doctype html>
<meta name=variant content="?dialog">
<meta name=variant content="?CloseWatcher">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/common/top-layer.js"></script>
<script src="../resources/helpers.js"></script>

<body>
<script>
const type = location.search.substring(1);

promise_test(async t => {
  const events = [];

  const watcher1 = await createBlessedRecordingCloseWatcher(t, events, "watcher1", type);
  await createBlessedRecordingCloseWatcher(t, events, "watcher2", type, watcher1);

  await sendCloseRequest();
  await waitForPotentialCloseEvent();
  assert_array_equals(events, ["watcher2 cancel[cancelable=true]", "watcher2 close"]);

  await sendCloseRequest();
  await waitForPotentialCloseEvent();
  assert_array_equals(events, ["watcher2 cancel[cancelable=true]", "watcher2 close", "watcher1 cancel[cancelable=true]", "watcher1 close"]);
}, "Create two close watchers with user activation");
</script>
